fix : type 분리
