+++ admin/src/apis/login/api.ts
@@ -1,11 +1,11 @@
 import { BASE_URL, HEADERS } from '@/constants/api';
 import { LoginRequest } from '@/type/login/type';
 
-export const postLogin = async ({ account, password }: LoginRequest) => {
+export const postLogin = async (body: LoginRequest) => {
   const res = await fetch(`${BASE_URL}/login`, {
     method: 'POST',
     headers: HEADERS,
-    body: JSON.stringify({ account: account, password: password }),
+    body: JSON.stringify(body),
   });
 
   return res.json();
