+++ admin/src/apis/main/api.ts
@@ -1,22 +1,17 @@
-import { BASE_URL, HEADERS } from '@/constants/api';
-import { getCookie } from '@/utils/cookie';
+import { AuthorizationHeader, BASE_URL } from '@/constants/api';
 
 export const getEventsData = async () => {
-  const accessToken = getCookie('accessToken');
-  const headers = { ...HEADERS, Authorization: `Bearer ${accessToken}` };
   const response = await fetch(`${BASE_URL}/event`, {
     method: 'GET',
-    headers: headers,
+    headers: AuthorizationHeader,
   });
   return response.json();
 };
 
 export const getWinnerData = async () => {
-  const accessToken = getCookie('accessToken');
-  const headers = { ...HEADERS, Authorization: `Bearer ${accessToken}` };
   const response = await fetch(`${BASE_URL}/winner`, {
     method: 'GET',
-    headers: headers,
+    headers: AuthorizationHeader,
   });
 
   return response.json();
