+++ admin/src/type/winManagement/type.ts
@@ -9,3 +9,9 @@ export interface GetFCFSWinnerRequest {
 export interface PostFCFSWinnerRequest {
   fcfsWinnerNum: number;
 }
+
+export interface PostRaffleWinnerRequest {
+  firstWinnerNum: number;
+  secondWinnerNum: number;
+  thirdWinnerNum: number;
+}
